# packages/contracts/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/audit/events": {
      "get": {
        "operationId": "AuditController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEventDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "audit"
        ]
      }
    },
    "/api/v1/auth/sessions": {
      "post": {
        "operationId": "AuthController_createSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSessionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "operationId": "AuthController_currentSession",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionIdentityDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/auth/sessions/current/revoke": {
      "post": {
        "operationId": "AuthController_revokeSession",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/organizations/current": {
      "get": {
        "operationId": "OrganizationsController_getCurrent",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "organizations"
        ]
      }
    },
    "/api/v1/organizations/current/memberships": {
      "get": {
        "operationId": "OrganizationsController_listMemberships",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MembershipDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "organizations"
        ]
      },
      "post": {
        "operationId": "OrganizationsController_addMembership",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMembershipRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddMembershipResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "organizations"
        ]
      }
    },
    "/api/v1/health": {
      "get": {
        "operationId": "HealthController_liveness",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Verifica se o processo da API está ativo",
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/health/ready": {
      "get": {
        "operationId": "HealthController_readiness",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Verifica se a API e suas dependências estão prontas",
        "tags": [
          "health"
        ]
      }
    }
  },
  "info": {
    "title": "Sistema ERP API",
    "description": "Contrato HTTP público do Sistema ERP",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http"
      }
    },
    "schemas": {
      "ApiErrorBodyDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "RESOURCE_NOT_FOUND"
          },
          "message": {
            "type": "string",
            "example": "Recurso não encontrado"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          },
          "requestId": {
            "type": "string",
            "example": "req_019..."
          }
        },
        "required": [
          "code",
          "message",
          "details",
          "requestId"
        ]
      },
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBodyDto"
          }
        },
        "required": [
          "error"
        ]
      },
      "AuditEventDto": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string"
          },
          "actorUserId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "correlationId": {
            "type": "string"
          },
          "entityId": {
            "type": "string",
            "nullable": true
          },
          "entityType": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "action",
          "actorUserId",
          "correlationId",
          "entityId",
          "entityType",
          "id",
          "metadata",
          "occurredAt",
          "requestId"
        ]
      },
      "CreateSessionRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@example.test"
          },
          "organizationSlug": {
            "type": "string",
            "example": "demo"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        },
        "required": [
          "email",
          "organizationSlug",
          "password"
        ]
      },
      "CreateSessionResponseDto": {
        "type": "object",
        "properties": {
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "ADMIN",
              "MEMBER"
            ]
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "token": {
            "type": "string",
            "description": "Token opaco retornado somente na criação da sessão"
          }
        },
        "required": [
          "organizationId",
          "role",
          "userId",
          "expiresAt",
          "token"
        ]
      },
      "SessionIdentityDto": {
        "type": "object",
        "properties": {
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "ADMIN",
              "MEMBER"
            ]
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "organizationId",
          "role",
          "userId"
        ]
      },
      "OrganizationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "MembershipDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "ADMIN",
              "MEMBER"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "SUSPENDED"
            ]
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "email",
          "id",
          "name",
          "role",
          "status",
          "userId"
        ]
      },
      "AddMembershipRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "member@example.test"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MEMBER"
            ]
          }
        },
        "required": [
          "email",
          "role"
        ]
      },
      "AddMembershipResponseDto": {
        "type": "object",
        "properties": {
          "membership": {
            "$ref": "#/components/schemas/MembershipDto"
          },
          "replayed": {
            "type": "boolean",
            "description": "Indica reaproveitamento seguro da resposta anterior"
          }
        },
        "required": [
          "membership",
          "replayed"
        ]
      },
      "HealthResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok"
            ],
            "example": "ok"
          }
        },
        "required": [
          "status"
        ]
      },
      "ReadinessChecksDto": {
        "type": "object",
        "properties": {
          "database": {
            "type": "string",
            "enum": [
              "up"
            ],
            "example": "up"
          }
        },
        "required": [
          "database"
        ]
      },
      "ReadinessResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok"
            ],
            "example": "ok"
          },
          "checks": {
            "$ref": "#/components/schemas/ReadinessChecksDto"
          }
        },
        "required": [
          "status",
          "checks"
        ]
      }
    }
  }
}